requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
yt-dlp